requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
psycopg2